back_end/university_partnerships/requirements.txt
django
djangorestframework
django-solo